+++ piecrust/tests/commit.rs
@@ -14,7 +14,7 @@ fn read_write_session() -> Result<(), Error> {
 
     {
         let mut session = vm.genesis_session();
-        let id = session.deploy(module_bytecode!("counter"))?;
+        let id = session.deploy(module_bytecode!("counter"), None::<&()>)?;
 
         assert_eq!(session.query::<(), i64>(id, "read_value", &())?, 0xfc);
 
@@ -27,7 +27,7 @@ fn read_write_session() -> Result<(), Error> {
     // old counter value still accessible.
 
     let mut other_session = vm.genesis_session();
-    let id = other_session.deploy(module_bytecode!("counter"))?;
+    let id = other_session.deploy(module_bytecode!("counter"), None::<&()>)?;
 
     assert_eq!(other_session.query::<(), i64>(id, "read_value", &())?, 0xfc);
 
@@ -47,7 +47,7 @@ fn read_write_session() -> Result<(), Error> {
 fn commit_restore() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
     let mut session_1 = vm.genesis_session();
-    let id = session_1.deploy(module_bytecode!("counter"))?;
+    let id = session_1.deploy(module_bytecode!("counter"), None::<&()>)?;
     // commit 1
     assert_eq!(session_1.query::<(), i64>(id, "read_value", &())?, 0xfc);
     session_1.transact::<(), ()>(id, "increment", &())?;
@@ -77,8 +77,8 @@ fn commit_restore_two_modules_session() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
 
     let mut session = vm.genesis_session();
-    let id_1 = session.deploy(module_bytecode!("counter"))?;
-    let id_2 = session.deploy(module_bytecode!("box"))?;
+    let id_1 = session.deploy(module_bytecode!("counter"), None::<&()>)?;
+    let id_2 = session.deploy(module_bytecode!("box"), None::<&()>)?;
 
     session.transact::<(), ()>(id_1, "increment", &())?;
     session.transact::<i16, ()>(id_2, "set", &0x11)?;
@@ -117,7 +117,7 @@ fn multiple_commits() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
 
     let mut session = vm.genesis_session();
-    let id = session.deploy(module_bytecode!("counter"))?;
+    let id = session.deploy(module_bytecode!("counter"), None::<&()>)?;
     // commit 1
     assert_eq!(session.query::<(), i64>(id, "read_value", &())?, 0xfc);
     session.transact::<(), ()>(id, "increment", &())?;
@@ -158,7 +158,7 @@ fn concurrent_sessions() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
 
     let mut session = vm.genesis_session();
-    let counter = session.deploy(module_bytecode!("counter"))?;
+    let counter = session.deploy(module_bytecode!("counter"), None::<&()>)?;
 
     assert_eq!(session.query::<(), i64>(counter, "read_value", &())?, 0xfc);
 
@@ -244,7 +244,7 @@ fn squashing() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
 
     let mut session = vm.genesis_session();
-    let counter = session.deploy(module_bytecode!("counter"))?;
+    let counter = session.deploy(module_bytecode!("counter"), None::<&()>)?;
 
     assert_eq!(session.query::<(), i64>(counter, "read_value", &())?, 0xfc);