+++ piecrust/tests/everest.rs
@@ -12,7 +12,7 @@ pub fn height() -> Result<(), Error> {
 
     let mut session = vm.genesis_session();
 
-    let id = session.deploy(module_bytecode!("everest"))?;
+    let id = session.deploy(module_bytecode!("everest"), None::<&()>)?;
 
     for h in 0u64..1024 {
         session.set_meta("height", h);
@@ -26,7 +26,7 @@ pub fn height() -> Result<(), Error> {
 pub fn meta_data_optionality() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
     let mut session = vm.genesis_session();
-    let id = session.deploy(module_bytecode!("everest"))?;
+    let id = session.deploy(module_bytecode!("everest"), None::<&()>)?;
     let height: Option<u64> = session.transact(id, "get_height", &())?;
     assert!(height.is_none());
     Ok(())