+++ piecrust/tests/stack.rs
@@ -12,7 +12,7 @@ pub fn push_pop() -> Result<(), Error> {
 
     let mut session = vm.genesis_session();
 
-    let id = session.deploy(module_bytecode!("stack"))?;
+    let id = session.deploy(module_bytecode!("stack"), None::<&()>)?;
 
     let val = 42;
 
@@ -36,7 +36,7 @@ pub fn multi_push_pop() -> Result<(), Error> {
 
     let mut session = vm.genesis_session();
 
-    let id = session.deploy(module_bytecode!("stack"))?;
+    let id = session.deploy(module_bytecode!("stack"), None::<&()>)?;
 
     const N: i32 = 16;